ABC Escuela 6-Marzo-2026

+++ maestros/routes.py
@@ -27,9 +27,17 @@ def agregarMaes():
 
     return render_template("maestros/Maestros.html", form=create_form)
 
-@maestros.route("/detalle/<int:matricula>", methods=['GET'])
-def detallesMaes(matricula):
-    maes = Maestros.query.get_or_404(matricula)
+@maestros.route("/detallesMaes", methods=['GET'])
+def detallesMaes():
+
+    if request.method == 'GET':
+        matricula = request.args.get('matricula')
+        maes = db.session.query(Maestros).filter(Maestros.matricula == matricula).first()
+        nombre = maes.nombre
+        apellido = maes.apellido
+        especialidad = maes.especialidad
+        email = maes.email
+
     return render_template("maestros/detallesMaes.html", maes=maes)
 
 @maestros.route("/modificar/<int:matricula>", methods=['GET','POST'])
@@ -49,14 +57,33 @@ def modificarMaes(matricula):
 
     return render_template("maestros/modificarMaes.html", form=form)
 
-@maestros.route("/eliminar/<int:matricula>", methods=['GET'])
-def eliminarMaes(matricula):
-    maes = Maestros.query.get_or_404(matricula)
 
-    db.session.delete(maes)
-    db.session.commit()
+@maestros.route("/eliminarMaes", methods=['GET', 'POST'])
+def eliminarMaes():
+    create_form = forms.UserForm(request.form)
+    if request.method == 'GET':
+        matricula = request.args.get('matricula')
 
-    return redirect(url_for('maestros.listas_maestros'))
+        maes = db.session.query(Maestros).filter(Maestros.matricula == matricula).first()
+        create_form.id.data = maes.matricula
+        create_form.nombre.data = maes.nombre
+        create_form.apellido.data = maes.apellido
+        create_form.especialidad.data = maes.especialidad
+        create_form.email.data = maes.email
+
+    if request.method == 'POST':
+        matricula = create_form.id.data
+        maes = db.session.query(Maestros).filter(Maestros.matricula == matricula).first()
+        maes.nombre=create_form.nombre.data
+        maes.apellido=create_form.apellido.data
+        maes.especialidad=create_form.especialidad.data
+        maes.email=create_form.email.data
+        db.session.delete(maes)
+        db.session.commit()
+        return redirect(url_for('maestros.listas_maestros'))
+
+    return render_template("maestros/eliminarMaes.html", form=create_form)
+
 
 '''@maestros.route('/perfil/<nombre>')
 def perfil(nombre):
